uprev dependencies
uprev the key web libs so that anything new that gets built on this
will not be immediately out of date.

React change to v16 breaks a lot of stuff.  So left it at the
latest v15 release as of this time.

Bootstrap moved to sass flow from LESS on move to v4.  So left it
at the latest v3 release.